refactor(authflow): extract code/token issuance into an authcode package
Move sendCodeResponse and the per-response_type issue handlers (code, access
token, id_token) out of approval.go into an authflow/authcode package
(authcode.Issuer). The consent screen (handleApproval) and the consent check
stay as Handler orchestration; the Handler holds an *authcode.Issuer and
delegates issuing the response to it.
